+++ functions/src/api/auth.ts
